refactor: split lateral functions into a new submodule, cache metadata to improve I/O speed, revert results-changing bugfixes for another branch
