+++ src/monarchs/core/metadata.py
@@ -1,13 +1,17 @@
 """
 Run metadata for MONARCHS output files.
 
-This outputs model_setup parameters as JSON, with the intent of allowing
-the user to re-run a model simulation from this metadata in future -
-e.g. for if you change your model setup script after running. A
-future tool will be able to read this metadata and re-run the model with
-the relevant parameters. It also contains details of the exact model
-version used to generate a given output.
+This is written as netCDF global attributes on each output and
+checkpoint file. You can check it with ``ncdump -h``. This
+includes the code/dependency versions and git state, plus a ``model_setup``
+attribute holding the resolved configuration as a JSON string. The intent is
+to allow a run to be reconstructed from its output - e.g. if the model setup
+script is changed after running. A future tool will read this back and re-run
+the model with the relevant parameters.
 TODO - actually write this tool
+
+Note: metadata is only written when an output/checkpoint file is produced,
+i.e. with ``save_output`` and/or ``dump_data`` enabled.
 
 Additionally, we handle variable metadata here for the output netCDFs.
 This is a WIP - not all fields are covered at time of writing, but the idea
@@ -26,25 +30,39 @@ from importlib.metadata import PackageNotFoundError, version
 import numpy as np
 
 
+# only run this once, then cache rather than regenerate each time
+_cache = {}
+
+
 def global_attrs(model_setup=None):
     """
-    Build a dict of provenance attributes for a netCDF file, suitable for
+    Build a dict of metadata attributes for a netCDF file, suitable for
     ``Dataset.setncatts``. All values are strings.
     """
+    attrs = {"created_utc": datetime.now(timezone.utc).isoformat()}
+    attrs.update(_run_metadata(model_setup))
+    return attrs
+
+
+def _run_metadata(model_setup):
+    """Run the metadata attach process, loading from cache"""
+    cached = _cache.get(id(model_setup))
+    if cached is not None:
+        return cached
+
     attrs = {
-        "created_utc": datetime.now(timezone.utc).isoformat(),
         "hostname": platform.node(),
         "python_version": platform.python_version(),
     }
+    # if the package isnt installed, attach a placeholder rather than nothing
     for package in ("monarchs-ice", "numpy", "numba", "scipy", "netCDF4"):
         try:
             attrs[f"{package}_version"] = version(package)
         except PackageNotFoundError:
-            pass
+            attrs[f"{package}_version"] = "not installed"
 
-    # get the git description of the model versioning used. if running
-    # from the pypi install, then it will use the exact release version, else
-    # it will be related to the branch/checkout used.
+    # git description of the model version. From a pypi install this is the
+    # exact release, from a checkout it reflects the branch/commit state.
     try:
         attrs["git_describe"] = subprocess.check_output(
             ["git", "describe", "--tags", "--dirty", "--always"],
@@ -52,11 +70,14 @@ def global_attrs(model_setup=None):
             stderr=subprocess.DEVNULL,
             text=True,
         ).strip()
+    # if no git, then don't fail silently, just say git isnt a thing
     except (OSError, subprocess.CalledProcessError):
-        pass
+        attrs["git_describe"] = "unavailable"
 
     if model_setup is not None:
         attrs["model_setup"] = json.dumps(_summarise_config(model_setup))
+
+    _cache[id(model_setup)] = attrs
     return attrs
